Move cache to separate file
